AI components load and store an AI lua script and call the "update" function that should be defined in the script

+++ Src/World/Camera.cpp
@@ -30,7 +30,7 @@ namespace glz
 
 				Drawable *drawable = (Drawable*)(mWorld->mDrawableSystem->getComponentForEntity(spatial->getHost()));
 				Float rotation = (Float)spatial->getHeading().angle(Vec2d());
-				drawable->draw(screenX, screenY, rotation-(PI/2.0f));
+				drawable->draw(screenX, screenY, rotation-((Float)PI/2.0f));
 
 				++iter;
 			}
